+++ previous-versions/commerce/src/sdk/models/shared/errormessage.ts
@@ -8,7 +8,7 @@ import { Expose } from "class-transformer";
 /**
  * Your `query` parameter was not correctly formed
  */
-export class Schema extends SpeakeasyBase {
+export class ErrorMessage extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "canBeRetried" })
     canBeRetried?: string;